requirements.txt
boto3==1.39.16
python-dotenv==1.1.1
openai==1.100.1
tenacity==9.1.2
httpx==0.28.1
deepgram-sdk==4.8.1
yt-dlp==2025.7.21